pyproject.toml
[tool.poetry]
name = "blow-with-django"
version = "0.1.0"
description = ""
authors = ["Nayan Biswas <nayan32biswas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.2.0"
djangorestframework-simplejwt = "^5.3.0"
typing-inspect = "0.8.0"
typing-extensions = "4.5.0"
mypy = "^1.5.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
drf-spectacular = "^0.26.5"
coverage = "^7.3.2"
markdown = "^3.5"
black = "^23.9.1"
flake8 = "^6.1.0"
isort = "^5.12.0"
pre-commit = "^3.4.0"
ipython = "^8.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
